Update: separate password and email change to two components Remove: EditAccountSettings file
